Profile page data retrieveal improved.
Session improved to contain all necessary user data in accordance with privacy options.

+++ src/main/java/bucoop/dto/UseradditionalDto.java
@@ -7,6 +7,6 @@ public interface UseradditionalDto {
     
     void persistUserAdditionalInfo(Useradditional userAdditional);
     
-    Useradditional getUserAdditionalWithId(long userId);
+    Useradditional getUserAdditionalWithId(Integer userId);
     
 }
