+++ src/main/java/bucoop/dto/UserbasicDto.java
@@ -10,7 +10,7 @@ public interface UserbasicDto {
 
     Userbasic getAppUserWithName(String username);
 
-    Userbasic getAppUserWithId(long userId);
+    Userbasic getAppUserWithId(Integer userId);
 
     void setType(String username , String userType);
 }
