+++ src/main/java/bucoop/dto/UserbasicDtoImpl.java
@@ -52,7 +52,7 @@ public class UserbasicDtoImpl implements UserbasicDto {
     }
 
     @Override
-    public Userbasic getAppUserWithId(long userId) {
+    public Userbasic getAppUserWithId(Integer userId) {
         final Userbasic userBasic = entityManager.find(Userbasic.class, userId);
         if (userBasic == null) {
             return null;
